bot/requirements.txt
discord.py>=2.3.0
httpx>=0.27.0
aiohttp>=3.9.0
